Move OSDM stuff into places controller
