+++ web/src/controllers/mod.rs
@@ -1,2 +1,1 @@
-pub mod osdm;
 pub mod places;
