+++ web/src/controllers/places.rs
@@ -1,4 +1,3 @@
-use crate::controllers::osdm::*;
 use crate::db::error::DbError;
 use crate::db::finders::*;
 use crate::db::station_record::StationRecord;
@@ -6,7 +5,63 @@ use crate::state::SharedAppState;
 use axum::extract::{Path, State};
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Json, Response};
+use serde::{Deserialize, Serialize};
 use std::convert::From;
+
+#[derive(Deserialize, Serialize, Debug, Clone, PartialEq)]
+pub struct OsdmGeoPosition {
+    pub latitude: f64,
+    pub longitude: f64,
+}
+
+#[derive(Deserialize, Serialize)]
+pub struct OsdmLink {
+    rel: String,
+    href: String,
+    _type: String,
+    value: String,
+}
+
+#[derive(Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OsdmPlace {
+    pub id: String,
+    pub object_type: String,
+    pub name: String,
+    pub alternative_ids: Vec<String>,
+    pub geo_position: Option<OsdmGeoPosition>,
+    pub _links: Vec<OsdmLink>,
+}
+
+//
+// Requests
+//
+
+#[derive(Deserialize, Serialize)]
+pub struct OsdmInitialPlaceInput {
+    pub name: Option<String>,
+}
+
+#[derive(Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OsdmPlaceRequest {
+    pub place_input: Option<OsdmInitialPlaceInput>,
+}
+
+//
+// Responses
+//
+
+#[derive(Deserialize, Serialize)]
+pub struct OsdmPlaceResponse {
+    pub places: Vec<OsdmPlace>,
+}
+
+#[derive(Deserialize, Serialize)]
+pub struct OsdmProblem {
+    pub code: String,
+    pub title: String,
+}
 
 pub enum PlacesResponse {
     Ok(OsdmPlaceResponse),
